Check if we have a network before making api calls

+++ app/src/main/java/com/amsterdam/marktbureau/makkelijkemarkt/api/ApiDeleteDagvergunning.java
@@ -43,15 +43,20 @@ public class ApiDeleteDagvergunning extends ApiCall {
      * @param callback the object that will process the api response
      */
     @Override
-    public void enqueue(Callback callback) {
-        super.enqueue(callback);
+    public boolean enqueue(Callback callback) {
+        if (super.enqueue(callback)) {
 
-        // if we have a payload we set the function to call and enqueue the async request
-        if (callback != null && mId != null) {
-            Call<String> call = mMakkelijkeMarktApi.deleteDagvergunning(mId);
+            // if we have a payload we set the function to call and enqueue the async request
+            if (callback != null && mId != null) {
+                Call<String> call = mMakkelijkeMarktApi.deleteDagvergunning(mId);
 
-            // call the api asynchronously
-            call.enqueue(callback);
+                // call the api asynchronously
+                call.enqueue(callback);
+
+                return true;
+            }
         }
+
+        return false;
     }
 }
